Move helper methods into a test trait so it can be used for both functional and kernel.
